tutorial #23

+++ app/Http/Controllers/IdeaController.php
@@ -18,11 +18,19 @@ class IdeaController extends Controller
             'content' => 'required|min:10|max:255' //validation rule
         ]);
 
+        $validated['user_id'] = auth()->id();
+
         Idea::create($validated);
         return redirect()->route('dashboard')->with('success','Idea is created successfully!'); //redirect to dashboard route page
     }
 
+
+
     public function destroy(Idea $idea){
+
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        };
 
         $idea->delete();
 
@@ -35,12 +43,20 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea){
 
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        };
+
         $editing = true;
 
         return view("ideas.show",compact('idea','editing'));
     }
 
     public function update(Idea $idea){
+
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        };
 
         $validated = request()->validate([
             'content' => 'required|min:10|max:255'
